Improve cmd file play
This supports the playing, and saving(making a snapshot file) of game files. These files can be used to demonstrate, test, and verify play which has been saved in a human readable form on files.  The control/execution of these features is presented in the File Control window, which can be brought up via the play window (crs_dots) CmdFile menu button.  Playable files can be found in the csrc folder.  Playable snapshot files may be found in the gm_snaps folder.